Fix race condition with tspClient and experiment
When opening the webview, there can be a race condition when the
tspClient and/or experiment were sent to the webview before the webview
was ready to receive it. The webview now notices the extension of its
ready state.

Also add react-components dependencies as they cannot be picked up
because of the manual dependency link

Signed-off-by: Geneviève Bastien <gbastien@versatic.net>